modified:   client/Client.java 	modified:   client/Config.java 	modified:   client/cmdset/CommandSet.java 	new file:   client/cmdset/executor/GetMessageExecutor.java 	modified:   client/cmdset/executor/SendBroadcastExecutor.java 	modified:   client/cmdset/executor/SendMessageExecutor.java 	new file:   client/cmdset/executor/params/CountQueuesMode.java 	new file:   client/cmdset/executor/params/SizeMessageMode.java 	modified:   client/read/TaskRead.java 	modified:   resources/client/config.properties 	modified:   resources/client/inputCommands.txt 	modified:   server/sr/dao/ClientDaoImpl.java 	new file:   server/sr/task/GetClientListTask.java 	modified:   server/sr/task/GetMessageTask.java 	modified:   server/sr/task/ParseTask.java

+++ server/sr/task/GetMessageTask.java
@@ -41,7 +41,9 @@ public class GetMessageTask implements Runnable{
       synchronized (guard) {
         MessageDao messageDao = getMessageDao();
         message = messageDao.last(command.getQueueId(), command.getClientId());
-        messageDao.delete(message);
+        if(message != null){
+          messageDao.delete(message);
+        }
         endExecSqlTime = System.currentTimeMillis() - startExecTime;
       }
       if (message != null) {
